Split out debugging code for BoardManager
This is obviously not going to end up in production, so just split
these logic into a new Unity script that only exists in the test
scene.
